+++ app/Http/Controllers/GrindingController.php
@@ -113,17 +113,17 @@ class GrindingController extends Controller
             return back()->with('custom_errors', 'You don`t have Right Permission');
         }
         $pelletes = Pellete::select('id', 'pellete_no')->get();
+        $grinding = Grinding::find($id);
 
         $batchIdsWithoutGrinding = ProductionOrder::join('purchase_orders', 'purchase_orders.id', '=', 'production_orders.order_id')
-            ->whereNotIn('production_orders.batch_id', function ($query) use ($id) {
+            ->whereNotIn('production_orders.batch_id', function ($query) use ($grinding) {
                 $query->select('grindings.batch_id')->from('grindings')
                 ->whereNull('grindings.deleted_at')
-                ->where('grindings.batch_id', '!=', $id);
+                ->where('grindings.batch_id', '!=', $grinding->batch_id);
             })
             ->pluck('production_orders.batch_id');
         $products = Product::find('id', 'name');
         $batches = Batch::whereIn('id', $batchIdsWithoutGrinding)->get();
-        $grinding = Grinding::find($id);
         $shotblasts = GrindingPellete::where('gr_id', '=', $id)->get();
         $goods = GrindingGoodPellete::where('gr_id', '=', $id)->get();
         $rejects = GrindingRejectPellete::where('gr_id', '=', $id)->get();
@@ -197,6 +197,9 @@ class GrindingController extends Controller
     }
 
     public function destroy($id){
+        if (!Auth::user()->hasPermissionTo('Grinding Delete')) {
+            return back()->with('custom_errors', 'You don`t have Right Permission');
+        }
         $grinding = Grinding::find($id);
         GrindingPellete::where('gr_id', '=', $id)->delete();
         GrindingGoodPellete::where('gr_id', '=', $id)->delete();
@@ -216,10 +219,10 @@ class GrindingController extends Controller
             $shotblast = Shotblast::where('batch_id', $request->id)->first();
             $grindings = Grinding::where('batch_id', $request->id)->get();
 
-            $query = 'SELECT DISTINCT pelletes.id, pelletes.pellete_no, shotblast_details.weight, shotblast_details.pcs FROM shotblast_details INNER JOIN pelletes ON shotblast_details.pellete_id = pelletes.id WHERE sb_id =' . $shotblast->id;
+            $query = 'SELECT DISTINCT pelletes.id, pelletes.pellete_no, shotblast_details.weight, shotblast_details.pcs FROM shotblast_details INNER JOIN pelletes ON shotblast_details.pellete_id = pelletes.id WHERE sb_id =' . $shotblast->id . ' AND shotblast_details.deleted_at IS NULL';
             
             foreach ($grindings as $grinding) {
-                $query .= ' UNION SELECT DISTINCT pelletes.id, pelletes.pellete_no, grinding_good_pelletes.weight, grinding_good_pelletes.pcs FROM grinding_good_pelletes INNER JOIN pelletes ON grinding_good_pelletes.pellete_id = pelletes.id WHERE gr_id =' . $grinding->id . ' AND grinding_good_pelletes.deleted_at IS NULL AND shotblast_details.deleted_at IS NULL';
+                $query .= ' UNION SELECT DISTINCT pelletes.id, pelletes.pellete_no, grinding_good_pelletes.weight, grinding_good_pelletes.pcs FROM grinding_good_pelletes INNER JOIN pelletes ON grinding_good_pelletes.pellete_id = pelletes.id WHERE gr_id =' . $grinding->id . ' AND grinding_good_pelletes.deleted_at IS NULL';
             }
             
             $production = ProductionOrder::where('batch_id', '=', $request->id)->with('product', 'purchaseOrder')->latest()->first();
